Added soft deletes to Events.php Changed deleteallAssociatedEvents() in Forms.php to deleteAllFutureEvents() Changed FormsController.php so that the form only saves if there have been changes to the form fields
TODO elementary and secondary inspections won't save properly if edited. Form freezes

+++ app/Forms.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StoreForm;
 use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Str;
 
 class Forms extends Model
@@ -76,6 +77,7 @@ class Forms extends Model
                     'options' => $request->options[$key],
                     'help' => $request->help[$key],
                 ]);
+
             }
         }
 
@@ -115,6 +117,9 @@ class Forms extends Model
 
         // loop through each field in the request and compare to the section's fields
         if (isset($request->field_id)) {
+            $i = sizeof($request->label);
+            Log::debug('Max array ' . $request->label[$i-1]);
+            Log::debug('Size ' . $i);
             foreach ($request->field_id as $key => $value) {
 
                 // search database for field with unique name
@@ -122,6 +127,7 @@ class Forms extends Model
 
                 // if the field exists, update it
                 if ($field) {
+                    Log::debug('Updated field: ' . $request->label[$key]);
                     $field->update([
                         'sections_id' => $section_ids[$request->section_id[$key]],
                         'label' => $request->label[$key],
@@ -132,7 +138,9 @@ class Forms extends Model
                     ]);
 
                     $field->save();
+
                 } else {
+                    Log::debug('Created field: ' . $request->label[$key]);
                     // if the field is new, add it to the database
                     Fields::create([
                         'sections_id' => $section_ids[$request->section_id[$key]],
@@ -143,6 +151,7 @@ class Forms extends Model
                         'options' => $request->options[$key],
                         'help' => $request->help[$key],
                     ]);
+
                 }
             }
         }
@@ -198,11 +207,13 @@ class Forms extends Model
         else return null;
     }
 
-    public function deleteAllAssociatedEvents()
+    public function deleteAllFutureEvents()
     {
         $events = $this->events;
         foreach ($events as $event) {
-            Events::destroy($event->id);
+            if ($event->date > Carbon::today()) {
+                Events::destroy($event->id);
+            }
         }
     }
 }
